Implement a mouse-following cursor
I need to be able to translate between screen coordinates and board
coordinates in order to present a coherant user interface. I've
demonstrated the correctness of my translation logic with a
grass-textured hexagon that follows the mouse as it is dragged across
the game board.

I was actually super-impressed with the usefulness of unit testing to
implement the translation method.