Re-add Hex as a component
I removed the Hex registration line at some point and it caused me some
grief when I was trying to do a Component lookup for them and couldn't
find any Hexes at all in the World.

I don't really know how Hex sits in the ECS paradigm, but I'm going to
leave it in for now until it becomes clearer to me.