Fix heuristic bug
I was using "1" as the value for the cost of moving one hex, but then
adding it to the heuristic values. I wasn't intending these values to be
compared to anything other than themselves, as their values are an
implementation detail (currently the square of the hypotenuse).

I've refactored to use the same heuristic function to calculate the cost
of moving from 0,0 to 0,1  as the "1" value, so that this is more
resilient in future.