Replace Impassable Hexes with Obstacle Component
I was getting stuck trying to understand what a "Hex" is. I had hacked
in a quick Impassable flag that was working, but the Hex concept was
being used in odd contexts.

I think splitting out an "Obstacle" as a Component is a good idea,
because it allows Actors to manipulate their own Obstacle to prevent
other Actors from moving to their Hex.