Replace faiface/pixel with ebiten
I tried ebiten as an experiment, and found I preferred the interface.

There are a few small changes all over the place, but the biggest change
here is the removal of Colors from Sprite.

I was not able to replicate how faiface tints sprites in ebiten, and
after a lot of struggling, I actually decided to remove the
functionality for now until I wanted it. The only usage was for the
cursor, and the semi-transparency will actually pollute the color
palette. Might intentionally leave this out.