Expand the control architecture
I started out with the idea that doing everything in main.go was not
going to hold up much long, and wanted to separate out some
responsibilities into different places.

What I ended up on here comes from thinking that

- main should be responsible for translating raw input like IsMouseDown
and MouseMove to game-specific inputs like "Interaction(x,y)"

- combat should be responsible for deciding what an interaction means in
the context of an encounter. It might be a click on a skill that the
player wants to cast, or it might be a click on a hex that the player
wants to cast their spell _on_, or move to.

- The intent system is responsible for translating a Movement intent to
a navigation from Hex A to Hex B.