Unify Navigate, Navigate4, Navigate7 implementations
The headline is nice on its own, but it's not why I did it. The goal
here was to refactor Navigation so that Action Point costs could be
surfaced outside the Navigation logic. I found it a bit tedious to
implement this in three places, so the refactor came about.