Add a bitmap font
The next thing I need to get the HUD going is the ability to render
text.

I did some research on fonts and found some nice ones, but I also got
very confused by copyright law. In the USA, typefaces cannot be
copyrighted, but _fonts_ can, but in Germany, typefaces are subject to
copyright.

I've been toying around with the idea of doing my own font
with a bitmap atlas since I had issues with SFML font rendering being so
different on different platforms. I decided to do do this now and I'm
pretty happy with the result.

The design I came up with I tried to infuse with the feeling of really
cringey 80's fantasy with some curls and angles. I didn't trace any
other typeface for this, it was all hand-placed pixels.