Move Events out of event package
Having the definitions for these game events in a shared package was
preventing me from including types in those events from the packages
that consumed them.

I think having events types all around the codebase is the better way; I
saw the event package as a generic one when I was designing it, but
lumped some implementations in there anyway for simplicity.