Add a wait for time to pass icon
When the game is in the preparing state, then it appears unresponsive,
so I added in an icon that indicates that we are waiting for something
significant to occur.

This also necessitated some extension to the way screen size is done in
game, which forced me to hoist the event.Bus up to main!