Move current actor name off portrait
It was hard to read the name of the Actor. I considered adding a
dropshadow effect, but decided to shrink the current stats section
instead, so that the name could fit at the top. I think the smaller font
for the stats actually works realy well, because regular font was a
little too prominent for something that should be pretty subtle.