Refactor Navigation dependencies
in 9dde44231c8abf0b9ceb7a597eb732292da54bfb I refactored Navigation to
accept functions that provided the answers for "is there a hex at M,N?"
and "what will it costs to move to M,N?".

The functions were declared inline inside the intentSystem's Update
method, which made them unable to be shared. I want to be able to
navigate in the context of showing a trail of cursors to the player to
show where the actor will go when they click on a hex, so these
functions need to be reusable now.