Add a "Hidden" Component
I want a way to retain the existence of an Entity, but omit it from
rendering so that I can work on different game modes, like overworld vs
combat. The use case I have is to keep all the state of the overworld in
memory, but not render any of it while a cobat is in progress.