Add SceneWipeSystem
I wanted a way to hide and reveal the screen without fading to black, as
my intent for this game is to only use a limited number of colors, and
alpha blending would generate many different shades.

This commit includes a matrix scene wipe that is oriented diagonally.

I believe the way that the tiles are picked is not 100% correct, but the
effect is pretty decent, so I'm leaving it as is.