Refactor combat.Manager responsibilities
I got so lost here, but after swapping stuff around, it feels like I'm
at a good place to call it and commit something. This is a giant blob of
a commit that does a heap of stuff, but I think the core idea I
implemented here was to have the combat.Manager be responsible for
creating new combat.Participant Entities when you call Begin, and
destroying those Entities when you call End.