Make DiagonalMatrixWipe obscure loading entities
I noticed that the scene would pop into view for a frame, then disappear
when the DiagonalMatrixWipe kicked in. I thought about why this would
happen, and realised that Begin loaded stuff immediately, but the
DiagonalMatrixWipes would not load their own pieces until the scene wipe
system swings around on the next frame. I think moving the Begin() code
into a callback on the scene wipe works pretty well to fix this.