Add enemy squad variety
This is a rough, hard-coded version of what I want to do - have
something like an "overworld recipe" that configures bounds for where
the paths can travel, terrain and skin/visual configuration for the
different hexes, and a "spawn table", of baddies, travelling merchants,
chests, random events etc that can appear in the world and how likely
they are.

When I get to that point, I would need a more generic way of "booting
up" an instance of an enemy, but for now this will give me the variety
I'm looking for.