overworld: Add recipe parsing
This adds some recipe variation from an external squads archive file
(not included here). That was my main goal, but I'm also rethinking how
I'm including the artwork for this game, as the source code here is
GPLed, and I think I would like to distribute the artwork with a
different (personal use?) license. This commit proves out a way of
loading game data from a file.