overworld: Add permissable paths
Prior to this commit, the connection between two hexes was determined
only by whether they were adjacent or not. This commit adds Paths to
overworld recipes which ought to become the source of truth for whetehr
a path can exist between two overworld hexes. This is necessary so that
terrain that includes embankments can be used.

N.B this only adds the values to the recipe, overworld map generation
still uses the old rules!