game/combat: Add a damage system
I need something to handle the calculations of damage being sent to a
target, and then that damage being reduced by the the target's defenses.
I'm reasonably confident in this implementation, but parts of it are
sketched out due to not having concrete types to use for resistances yet
(defense to reduce physical damage, or fire, ice, etc resistances).