combat: Allow virtual trigger points for skill effects
I've wanted to add this in for a while - a skill effect should be able
to be triggered at the natural point of the user's attack or spell
animation, whatever that may be. Otherwise, all skillEffects would need
to be duplicated for each profession that gets it, and weapon-based
skills would be impossible to implement other than having every
profession use then at the same time.

This refactor has also happily decoupled a skill effect from its timing.
No more awkward Effect interface and duplicated Schedule method!