renderer: Move entity rendering to its own func
The diff is pretty unreadable here, but basically all I did was move the
bulk of the implementation of Render() into a new method called
renderEntity(). It doesnt get a reference to the screen parameter, so
mistakes where you render directly instead of via the offscreen are
harder to make.