combat: Make Obscurers disappear when they would obscure
I want to add tiles like trees to maps, but the foliage would block out
the view of the game board pretty significantly. My take on resolving
this is to make the tree sprites disappear when something goes behind
them. I can keep a second sprite that's only an outline that does not
disappear, and maintain the presence of the tree, while giving the
effect of transparency, without actually using any alpha blending.