Fix a layout bug
The issue here was that the calculation of the widest child of an
element forgot to account for sequential columns needing to have their
width's summed, instead of just using whichever is widest.

There's a complicated way to work this out - by summing the column widths of
every unbroken sequence of columns, but resetting for every new column
sequence. I've done the cheeky way though, because the assumption being
made elsewhere in the code is that columns are always x/12th of the
total width, and every sequence of columns is 12/12ths of the total
width.