ui: Invent an If element
So it turns out these UI declarations will need a bit more logic in
them.

The use case I have for <If> is the Turn queue in the combat mode.
There's an uncertain number of participants, and a fixed number of
slots, so if there are fewer participants than slots, the excess slots
must not be rendered.