Wander through ui code, refactoring
Apologies to any readers. This is a messy commit.

It ...

- Starts on a combat hud UI to replace the existing imperative one.

- It merges all the ui code into one place that gets called every
  update.  This allows handling clicks and rendering a ui to be simple,
because all the info is readily available. Was getting so stuck having
to implement things in several places, and worrying if bugs were in the
rendering code or the click handling code when buttons didn't work.

- Adds <Range />, stubs out some of <For />

- Removes Twelfthsoffsets as a concept. It was impossible to calculate
  the correct values when a <Column /> as a child of a <Range /> could
not know the number of its siblings until data had been applied.

- Codifies that children must be exclusively <Column />s or exclusively
  non-<Column />s.