ui: Refactor one more time, with feeling
I was despairing to add the logic to handle clickable icons and address
the bugs I was finding; the way calculating children worked felt like it
was all corner cases. I struggled with this refactor a lot, but I'm
very happy with how it came out. The code is far easier to follow now,
and there's less of it. Huge lines deleted score here.