ui: Add ids to interactives
I want to be able to define a single handler for skills which accepts
the id of the clicked skill. I've extended the concept of buttons and
clickable images to have an id which identifies which button is pressed,
so that UI authors can optionally reuse interaction handlers by passing
context through to different visible callers.