Fix click through UIs bug
This one specifically fixes embark mode where when you click on a button
of the focused villager modal that happens to be panned over the top of
one of the villagers in the world, both interactive regions are
triggered.

I've sequenced this so that all UI interactions are handled, and only
then if a UI did not handle the interaction, is the event passed to the
regular interactives.