tools/cmd/overworld: Build some better output config
Adding the ability to force specific seeds and recipes so that I can
focus on a bug I stumbled on with the dark forest recipe where it
generated a bit of path that wasn't connected to the start of goal.

This is also going to allow the attract mode that randomly picks a
recipe each time you ctrl+R to reload.