Refactor archive loading
I want to reuse the core interpretation logic in more than this context,
and the original implementation entangled that with code to read tar
files. I noticed some odd things in this flow (ie only reading pngs if
they're in a certain folder), and wanted to simplify some of those
checks as well. Hopefully I didn't break anything!