Make ui Buttons animated on click
I felt like the UI is a bit unpolished, and wanted to make it so that
the buttons have more interactivity. They should look different when you
press them etc. I had difficulties when thinking about how to make them
look pressed when you click down, and trigger when you release, so I
ended up faking that by introducing a "delayer" that will visually show
the button pressed for a few milliseconds, and then trigger the handler.

I feel like this code is unwieldy. Take a look at the args to
calculateChildren for example. Yikes. This works but it's ugly.